Minor Release: v0.1.2
Minor release to main branch. Updated the redirect link for the 'Apply Now' button on the index.html page. Added a new customer review to the about-us.html page. Fixed typos and added new heading to the about-us.html page.

The full changelog for all commits/versions is located in the README.md file within the main repository.