Minor Release: v0.1.4
Minor release to main branch. Several code refactoring and tweaks were made to the styles.css and several .html files. A custom favicon was also added to each current webpage. A new pull request for a separate exp branch was also made for testing and storing experimental code.

The full changelog for all commits/versions is located in the README.md file within the main repository.